+++ basic/chat/Server.java
@@ -1,5 +1,5 @@
 
-package basic;
+package chat;
 
 import java.io.IOException;
 import java.io.ObjectInputStream;
